Fix Regexp used for CSS comment detection (#32)
* Fix `SOURCE_MAPPING_PATTERN` to correctly match CSS comments

Previously the `*` was not escaped in this Regexp, so it was being
interpreted as a "zero or more" quantifier on the previous character,
rather than as a literal `*` as intended. This meant that the Regexp
matched the prefix `# sourceMappingURL=` even when it did not appear
inside a CSS comment. Now it'll only match `/*# sourceMappingURL=`.

* Fix inline Regexp in `SourceMappingUrlsTest`

Again, the intent here was to match the beginning of a CSS comment
string (i.e. to match the literal prefix `/*`), but this was actually
matching "zero or more" `/` characters because the `*` was not escaped
in the Regexp.

* Construct Regexp literals using `%r` syntax

This way we don't have to escape literal `/` characters, which makes the
Regexp easier to follow.

* Wrap really long lines in `SourceMappingUrlsTest`

* Restrict `SOURCE_MAPPING_PATTERN` to only match leading comments

The `sourceMappingURL` comments always appear at the beginning of the
line, so we can ignore any comments that appear later in a line when
searching for comments to replace.

* Update `SOURCE_MAPPING_PATTERN` to assume there is at least 1 character

before the `.map` extension (i.e. the Regexp doesn't need to match
`/*# sourceMappingURL=.map`)